Memoized editing distance algorithm

+++ C/include/editing_distance.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int editing_distance(const char* string1, const char* string2);
+unsigned long editing_distance(const char* string1, const char* string2);
